+++ src/groups/permissions.rs
@@ -64,7 +64,7 @@ pub struct GroupOpenCloudPermissions {
 #[derive(Debug, Serialize, Deserialize, Clone, Type)]
 #[serde(rename_all = "camelCase")]
 pub struct RolePermissions {
-    pub group_id: u32,
+    pub group_id: i64,
     pub role: GroupRole,
     pub permissions: GroupPermissions,
 }
@@ -77,8 +77,8 @@ pub struct RolePermissions {
 /// - 3: You are not authorized to view/edit permissions for this role.
 pub async fn role_permissions(
     jar: &RequestJar,
-    group_id: u32,
-    role_id: u32,
+    group_id: i64,
+    role_id: i64,
 ) -> Result<RolePermissions, Box<Error>> {
     let url = format!(
         "https://groups.roblox.com/v1/groups/{}/roles/{}/permissions",
@@ -130,8 +130,8 @@ pub struct UpdateRolePermissionsRequest {
 /// - 4: This role's permissions can not be modified.
 pub async fn update_role_permissions(
     jar: &RequestJar,
-    group_id: u32,
-    role_id: u32,
+    group_id: i64,
+    role_id: i64,
     permissions: UpdateRolePermissionsRequestPermissions,
 ) -> Result<(), Box<Error>> {
     let url = format!(
@@ -153,7 +153,7 @@ pub async fn update_role_permissions(
 /// - 3: You are not authorized to view/edit permissions for this role.
 pub async fn guest_permissions(
     jar: &RequestJar,
-    group_id: u32,
+    group_id: i64,
 ) -> Result<RolePermissions, Box<Error>> {
     let url = format!(
         "https://groups.roblox.com/v1/groups/{}/roles/guest/permissions",
@@ -173,7 +173,7 @@ pub async fn guest_permissions(
 /// *Note: None were provided in the documentation*
 pub async fn permissions(
     jar: &RequestJar,
-    group_id: u32,
+    group_id: i64,
 ) -> Result<Vec<RolePermissions>, Box<Error>> {
     let url = format!(
         "https://groups.roblox.com/v1/groups/{}/roles/permissions",
