+++ src/groups/relationships.rs
@@ -8,7 +8,7 @@ use super::Group;
 
 #[derive(Debug, Serialize, Deserialize, Clone, Type)]
 pub struct BatchRequest {
-    pub group_ids: Vec<u32>,
+    pub group_ids: Vec<i64>,
 }
 
 /// Batch management of group relationships (enemies and allies)
@@ -23,8 +23,8 @@ pub mod relationships {
     #[allow(unused)] // It appears to be a bug in the checker, this kinda fixes it.
     pub async fn decline(
         jar: &RequestJar,
-        group_id: u32,
-        group_ids: Vec<u32>,
+        group_id: i64,
+        group_ids: Vec<i64>,
         relationship_type: RelationshipType,
     ) -> Result<(), Box<Error>> {
         if relationship_type == RelationshipType::All {
@@ -48,8 +48,8 @@ pub mod relationships {
     #[allow(unused)] // It appears to be a bug in the checker, this kinda fixes it.
     pub async fn accept(
         jar: &RequestJar,
-        group_id: u32,
-        group_ids: Vec<u32>,
+        group_id: i64,
+        group_ids: Vec<i64>,
         relationship_type: RelationshipType,
     ) -> Result<(), Box<Error>> {
         if relationship_type == RelationshipType::All {
@@ -81,8 +81,8 @@ pub mod relationship {
     #[allow(unused)] // It appears to be a bug in the checker, this kinda fixes it.
     pub async fn decline(
         jar: &RequestJar,
-        group_id: u32,
-        relation_group_id: u32,
+        group_id: i64,
+        relation_group_id: i64,
         relationship_type: RelationshipType,
     ) -> Result<(), Box<Error>> {
         let url = format!(
@@ -102,8 +102,8 @@ pub mod relationship {
     #[allow(unused)] // It appears to be a bug in the checker, this kinda fixes it.
     pub async fn accept(
         jar: &RequestJar,
-        group_id: u32,
-        relation_group_id: u32,
+        group_id: i64,
+        relation_group_id: i64,
         relationship_type: RelationshipType,
     ) -> Result<(), Box<Error>> {
         let url = format!(
@@ -142,9 +142,9 @@ impl RelationshipType {
 #[derive(Debug, Serialize, Deserialize, Clone, Type)]
 #[serde(rename_all = "camelCase")]
 pub struct GroupRelationships {
-    pub group_id: u32,
+    pub group_id: i64,
     pub relationship_type: RelationshipType,
-    pub total_group_count: u32,
+    pub total_group_count: i64,
     #[serde(rename = "relatedGroups")]
     pub groups: Vec<Group>,
 }
@@ -158,7 +158,7 @@ pub struct GroupRelationships {
 #[async_recursion::async_recursion]
 pub async fn relationships(
     jar: &RequestJar,
-    group_id: u32,
+    group_id: i64,
     relationship_type: RelationshipType,
 ) -> Result<GroupRelationships, Box<Error>> {
     if relationship_type == RelationshipType::All {
@@ -188,7 +188,7 @@ pub async fn relationships(
 /// - 1: Group is invalid or does not exist.
 /// - 4: Group relationship type or request type is invalid.
 /// - 8: Invalid or missing pagination parameters
-pub async fn enemies(jar: &RequestJar, group_id: u32) -> Result<GroupRelationships, Box<Error>> {
+pub async fn enemies(jar: &RequestJar, group_id: i64) -> Result<GroupRelationships, Box<Error>> {
     relationships(jar, group_id, RelationshipType::Enemy).await
 }
 
@@ -198,7 +198,7 @@ pub async fn enemies(jar: &RequestJar, group_id: u32) -> Result<GroupRelationshi
 /// - 1: Group is invalid or does not exist.
 /// - 4: Group relationship type or request type is invalid.
 /// - 8: Invalid or missing pagination parameters
-pub async fn allies(jar: &RequestJar, group_id: u32) -> Result<GroupRelationships, Box<Error>> {
+pub async fn allies(jar: &RequestJar, group_id: i64) -> Result<GroupRelationships, Box<Error>> {
     relationships(jar, group_id, RelationshipType::Ally).await
 }
 
@@ -212,7 +212,7 @@ pub async fn allies(jar: &RequestJar, group_id: u32) -> Result<GroupRelationship
 #[async_recursion::async_recursion]
 pub async fn relationship_requests(
     jar: &RequestJar,
-    group_id: u32,
+    group_id: i64,
     relationship_type: RelationshipType,
 ) -> Result<GroupRelationships, Box<Error>> {
     if relationship_type == RelationshipType::All {
@@ -245,7 +245,7 @@ pub async fn relationship_requests(
 /// - 8: Invalid or missing pagination parameters
 pub async fn enemy_requests(
     jar: &RequestJar,
-    group_id: u32,
+    group_id: i64,
 ) -> Result<GroupRelationships, Box<Error>> {
     relationship_requests(jar, group_id, RelationshipType::Enemy).await
 }
@@ -259,7 +259,7 @@ pub async fn enemy_requests(
 /// - 8: Invalid or missing pagination parameters
 pub async fn ally_requests(
     jar: &RequestJar,
-    group_id: u32,
+    group_id: i64,
 ) -> Result<GroupRelationships, Box<Error>> {
     relationship_requests(jar, group_id, RelationshipType::Ally).await
 }
@@ -273,8 +273,8 @@ pub async fn ally_requests(
 /// - 11: Relationship does not exist.
 pub async fn remove(
     jar: &RequestJar,
-    group_id: u32,
-    target_group_id: u32,
+    group_id: i64,
+    target_group_id: i64,
     relationship_type: RelationshipType,
 ) -> Result<(), Box<Error>> {
     if relationship_type == RelationshipType::All {
@@ -300,8 +300,8 @@ pub async fn remove(
 /// - 11: Relationship does not exist.
 pub async fn remove_enemy(
     jar: &RequestJar,
-    group_id: u32,
-    target_group_id: u32,
+    group_id: i64,
+    target_group_id: i64,
 ) -> Result<(), Box<Error>> {
     remove(jar, group_id, target_group_id, RelationshipType::Enemy).await?;
     Ok(())
@@ -316,8 +316,8 @@ pub async fn remove_enemy(
 /// - 11: Relationship does not exist.
 pub async fn remove_ally(
     jar: &RequestJar,
-    group_id: u32,
-    target_group_id: u32,
+    group_id: i64,
+    target_group_id: i64,
 ) -> Result<(), Box<Error>> {
     remove(jar, group_id, target_group_id, RelationshipType::Ally).await?;
     Ok(())
@@ -337,8 +337,8 @@ pub async fn remove_ally(
 /// - 9: Insufficient permissions.
 pub async fn send_request(
     jar: &RequestJar,
-    group_id: u32,
-    target_group_id: u32,
+    group_id: i64,
+    target_group_id: i64,
     relationship_type: RelationshipType,
 ) -> Result<(), Box<Error>> {
     if relationship_type == RelationshipType::All {
@@ -369,8 +369,8 @@ pub async fn send_request(
 /// - 9: Insufficient permissions.
 pub async fn send_enemy_request(
     jar: &RequestJar,
-    group_id: u32,
-    target_group_id: u32,
+    group_id: i64,
+    target_group_id: i64,
 ) -> Result<(), Box<Error>> {
     send_request(jar, group_id, target_group_id, RelationshipType::Enemy).await?;
     Ok(())
@@ -390,8 +390,8 @@ pub async fn send_enemy_request(
 /// - 9: Insufficient permissions.
 pub async fn send_ally_request(
     jar: &RequestJar,
-    group_id: u32,
-    target_group_id: u32,
+    group_id: i64,
+    target_group_id: i64,
 ) -> Result<(), Box<Error>> {
     send_request(jar, group_id, target_group_id, RelationshipType::Ally).await?;
     Ok(())
