+++ src/groups/search.rs
@@ -6,10 +6,10 @@ use crate::util::{jar::RequestJar, paging::PageLimit, responses::DataWrapper, Er
 #[derive(Debug, Serialize, Deserialize, Clone, Type)]
 #[serde(rename_all = "camelCase")]
 pub struct SearchGroup {
-    pub id: u32,
+    pub id: i64,
     pub name: String,
     pub description: String,
-    pub member_count: u32,
+    pub member_count: i64,
     pub previous_name: Option<String>,
     pub public_entry_allowed: bool,
     pub created: String,
@@ -20,9 +20,9 @@ pub struct SearchGroup {
 #[derive(Debug, Serialize, Deserialize, Clone, Type)]
 #[serde(rename_all = "camelCase")]
 pub struct MinimalSearchGroup {
-    pub id: u32,
+    pub id: i64,
     pub name: String,
-    pub member_count: u32,
+    pub member_count: i64,
     pub has_verified_badge: bool,
 }
