+++ src/groups/wall.rs
@@ -13,7 +13,7 @@ use crate::{
 #[derive(Debug, Serialize, Deserialize, Clone, Type)]
 #[serde(rename_all = "camelCase")]
 pub struct WallPost {
-    pub id: u32,
+    pub id: i64,
     pub poster: MinimalGroupUser,
     pub body: String,
     pub created: String,
@@ -27,7 +27,7 @@ pub struct WallPost {
 /// - 2: You do not have permission to access this group wall.
 pub async fn wall(
     jar: &RequestJar,
-    group_id: u32,
+    group_id: i64,
     limit: PageLimit,
     sort_order: Option<SortOrder>,
 ) -> Result<Vec<WallPost>, Box<Error>> {
@@ -52,8 +52,8 @@ pub async fn wall(
 /// - 3: The group wall post id is invalid or does not exist.
 pub async fn delete_wall_post(
     jar: &RequestJar,
-    group_id: u32,
-    post_id: u32,
+    group_id: i64,
+    post_id: i64,
 ) -> Result<(), Box<Error>> {
     let url = format!(
         "https://groups.roblox.com/v1/groups/{}/wall/posts/{}",
@@ -71,8 +71,8 @@ pub async fn delete_wall_post(
 /// - 6: The user specified is invalid or does not exist.
 pub async fn delete_wall_posts_by_user(
     jar: &RequestJar,
-    group_id: u32,
-    user_id: u32,
+    group_id: i64,
+    user_id: i64,
 ) -> Result<(), Box<Error>> {
     let url = format!(
         "https://groups.roblox.com/v1/groups/{}/wall/users/{}/posts",
